% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{wcm_palettes}
\alias{wcm_palettes}
\title{List of primary and secondary WCM palettes}
\format{An object of class \code{list} of length 2.}
\usage{
wcm_palettes
}
\description{
Use \code{\link{wcm_palette}} to construct palettes of desired length.
}
\keyword{datasets}
