% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{wcm_palette}
\alias{wcm_palette}
\title{WCM palette generator}
\usage{
wcm_palette(palette_name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{n}{Number of colors desired. If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Continuous automatically interpolates between colors.}

\item{name}{Name of palette. Derived from \href{https://brand.weill.cornell.edu/brand-guidelines/color-palette}{WCM branding}.
Choices are: \code{primary} (default), \code{secondary}}
}
\value{
A vector of colors.
}
\description{
Primary and secondary color palettes of Weill Cornell Medicine.
}
\examples{
wcm_palette("primary")
wcm_palette("primary", n = 10, type = "continuous")
}
\keyword{colors}
